refactor: 리팩토링 지대로 하기
